+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/camera/data/MapboxNavigationViewportDataSource.kt
@@ -481,14 +481,10 @@ class MapboxNavigationViewportDataSource(
             routeProgress.currentLegProgress,
             routeProgress.currentLegProgress?.currentStepProgress
         ) { currentLegProgress, currentStepProgress ->
-            options.followingFrameOptions.run {
-                followingPitchProperty.fallback = getPitchFallbackFromRouteProgress(
-                    pitchNearManeuvers.enabled,
-                    pitchNearManeuvers.triggerDistanceFromManeuver,
-                    defaultPitch,
-                    currentStepProgress.distanceRemaining
-                )
-            }
+            followingPitchProperty.fallback = getPitchFallbackFromRouteProgress(
+                routeProgress,
+                options.followingFrameOptions
+            )
 
             options.followingFrameOptions.intersectionDensityCalculation.run {
                 pointsToFrameOnCurrentStep = getPointsToFrameOnCurrentStep(